file structure improvements and  minor cleanup
