+++ packages/plexus-core/src/types.ts
@@ -1,6 +1,11 @@
 import { _event } from './event'
-import { _runtime } from './runtime'
+import { _runtime } from './instance/runtime'
 
 export type PlexusRuntime = ReturnType<typeof _runtime>
 export type PlexusEvent = ReturnType<typeof _event>
 export type PlexusWatcher<V extends any = any> = (value: V) => void
+
+export type PlexusInternalWatcher<V extends any = any> = (
+	value: V,
+	from?: string
+) => void
